♻️ Split tests (#1631)
