+++ src/test/lib/utils/task.test.ts
@@ -20,14 +20,10 @@ import {
   compareByContestIdAndTaskId,
   removeTaskIndexFromTitle,
   getTaskTableHeaderName,
-  toChangeTextColorIfNeeds,
-  getTaskGradeLabel,
-  toChangeBorderColorIfNeeds,
 } from '$lib/utils/task';
 import type { WorkBookTaskBase } from '$lib/types/workbook';
 import { ContestType } from '$lib/types/contest';
 import { type TaskResult, type TaskResults } from '$lib/types/task';
-import { TaskGrade } from '@prisma/client';
 
 type TestCaseForTaskResults = {
   taskResults: TaskResults;
@@ -67,13 +63,6 @@ type TestCaseForTaskTableHeaderName = {
 };
 
 type TestCasesForTaskTableHeaderName = TestCaseForTaskTableHeaderName[];
-
-type TestCaseForTaskGrade = {
-  taskGrade: TaskGrade;
-  expected: string;
-};
-
-type TestCasesForTaskGrade = TestCaseForTaskGrade[];
 
 describe('Task', () => {
   describe('task url', () => {
@@ -695,98 +684,6 @@ describe('Task', () => {
         testCases,
         ({ title, taskTableIndex, expected }: TestCaseForNewTitle) => {
           expect(removeTaskIndexFromTitle(title, taskTableIndex)).toBe(expected);
-        },
-      );
-    });
-  });
-
-  describe('to change text color if needs', () => {
-    describe('when task grades from 11Q to 1Q are given ', () => {
-      const testCases: TestCasesForTaskGrade = [
-        { taskGrade: TaskGrade.Q11, expected: 'text-black' },
-        { taskGrade: TaskGrade.Q10, expected: 'text-black' },
-        { taskGrade: TaskGrade.Q6, expected: 'text-black' },
-        { taskGrade: TaskGrade.Q5, expected: 'text-black' },
-        { taskGrade: TaskGrade.Q2, expected: 'text-black' },
-        { taskGrade: TaskGrade.Q1, expected: 'text-black' },
-      ];
-
-      runTests(
-        'toChangeTextColorIfNeeds',
-        testCases,
-        ({ taskGrade, expected }: TestCaseForTaskGrade) => {
-          expect(toChangeTextColorIfNeeds(getTaskGradeLabel(taskGrade))).toBe(expected);
-        },
-      );
-    });
-
-    describe('when task grades from 1D to 5D are given', () => {
-      const testCases: TestCasesForTaskGrade = [
-        { taskGrade: TaskGrade.D1, expected: 'text-white' },
-        { taskGrade: TaskGrade.D2, expected: 'text-white' },
-        { taskGrade: TaskGrade.D3, expected: 'text-white' },
-        { taskGrade: TaskGrade.D4, expected: 'text-white' },
-        { taskGrade: TaskGrade.D5, expected: 'text-white' },
-      ];
-
-      runTests(
-        'toChangeTextColorIfNeeds',
-        testCases,
-        ({ taskGrade, expected }: TestCaseForTaskGrade) => {
-          expect(toChangeTextColorIfNeeds(getTaskGradeLabel(taskGrade))).toBe(expected);
-        },
-      );
-    });
-
-    describe('when task grades 6D is given', () => {
-      const testCases: TestCasesForTaskGrade = [
-        { taskGrade: TaskGrade.D6, expected: 'text-atcoder-bronze' },
-      ];
-
-      runTests(
-        'toChangeTextColorIfNeeds',
-        testCases,
-        ({ taskGrade, expected }: TestCaseForTaskGrade) => {
-          expect(toChangeTextColorIfNeeds(getTaskGradeLabel(taskGrade))).toBe(expected);
-        },
-      );
-    });
-  });
-
-  describe('to change border color if needs', () => {
-    describe('when task grades from 11Q to 5D are given', () => {
-      const testCases: TestCasesForTaskGrade = [
-        { taskGrade: TaskGrade.Q11, expected: 'border-white' },
-        { taskGrade: TaskGrade.Q10, expected: 'border-white' },
-        { taskGrade: TaskGrade.Q7, expected: 'border-white' },
-        { taskGrade: TaskGrade.Q6, expected: 'border-white' },
-        { taskGrade: TaskGrade.Q2, expected: 'border-white' },
-        { taskGrade: TaskGrade.Q1, expected: 'border-white' },
-        { taskGrade: TaskGrade.D1, expected: 'border-white' },
-        { taskGrade: TaskGrade.D2, expected: 'border-white' },
-        { taskGrade: TaskGrade.D4, expected: 'border-white' },
-        { taskGrade: TaskGrade.D5, expected: 'border-white' },
-      ];
-
-      runTests(
-        'toChangeBorderColorIfNeeds',
-        testCases,
-        ({ taskGrade, expected }: TestCaseForTaskGrade) => {
-          expect(toChangeBorderColorIfNeeds(getTaskGradeLabel(taskGrade))).toBe(expected);
-        },
-      );
-    });
-
-    describe('when task grade 6D is given ', () => {
-      const testCases: TestCasesForTaskGrade = [
-        { taskGrade: TaskGrade.D6, expected: 'border-atcoder-bronze' },
-      ];
-
-      runTests(
-        'toChangeBorderColorIfNeeds',
-        testCases,
-        ({ taskGrade, expected }: TestCaseForTaskGrade) => {
-          expect(toChangeBorderColorIfNeeds(getTaskGradeLabel(taskGrade))).toBe(expected);
         },
       );
     });
